requirements.txt
# Core framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.23
alembic>=1.12.1

# AI/LLM clients
openai>=1.3.0
anthropic>=0.7.0

# Google Workspace integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
google-cloud-storage>=2.10.0

# Task queue and background processing
celery>=5.3.4
redis>=5.0.1

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0
python-dateutil>=2.8.2

# Data processing
pandas>=2.1.3
numpy>=1.25.2

# Logging and monitoring
structlog>=23.2.0
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Email processing
aiosmtplib>=3.0.1
email-validator>=2.1.0

# File processing
python-docx>=1.1.0
PyPDF2>=3.0.1
openpyxl>=3.1.2

# Security
cryptography>=41.0.7
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0